MoviesList function updated and moved from MovieList folder
