+++ src/core/App/index.js
@@ -2,7 +2,7 @@ import { ThemeProvider } from "styled-components";
 import { Normalize } from "styled-normalize";
 import { GlobalStyle } from "./GlobalStyle";
 import { theme } from "./theme";
-import { MovieList } from "../../features/MoviesList";
+import { MoviesList } from "../../features/MoviesList";
 import { Navigation } from "../../common/Navigation";
 
 function App() {
@@ -11,7 +11,7 @@ function App() {
       <Normalize />
       <GlobalStyle />
       <Navigation />
-      <MovieList />
+      <MoviesList />
     </ThemeProvider>
   );
 }
