+++ src/core/store.js
@@ -1,6 +1,6 @@
 import createSagaMiddleware from "@redux-saga/core";
 import { configureStore } from "@reduxjs/toolkit";
-import popularMoviesReducer from "../features/MovieList/popularMoviesSlice";
+import popularMoviesReducer from "../features/MoviesList/popularMoviesSlice";
 import rootSaga from "./rootSaga";
 
 const sagaMiddleware = createSagaMiddleware();
